Add --config CLI option to load defaults from file
Since we use click chain mode for commands, required arguments must
still be specified on the command line even if the have a default in the
config file. To solve this problem you can now use "." as the argument
value to force it to be loaded from the config. For example:
  trl --config conf:conf . fqi .